+++ app/service/task.js
@@ -1,8 +1,8 @@
 module.exports = app => {
   class TaskService extends app.Service {
     async findAll(query, conditions) {
-      let model =  this.ctx.model.Task;
-      let builder =  model.find(conditions);
+      let model = this.ctx.model.Task;
+      let builder = model.find(conditions);
       if (query.select) {
         select = JSON.parse(query.select);
         builder = builder.select(select);
@@ -26,9 +26,9 @@ module.exports = app => {
     }
 
     async findById(query, id) {
-      let model =  this.ctx.model.Task;
+      let model = this.ctx.model.Task;
       let select = {};
-      let builder =  model.findById(id);
+      let builder = model.findById(id);
       if (query.select) {
         select = JSON.parse(query.select);
         builder = builder.select(select);
@@ -38,19 +38,19 @@ module.exports = app => {
     }
 
     async create(body) {
-      let model =  this.ctx.model.Task;
+      let model = this.ctx.model.Task;
       const result = await model.create(body);
       return result;
     }
 
     async delete(id) {
-      let model =  this.ctx.model.Task;
+      let model = this.ctx.model.Task;
       const result = await model.findByIdAndRemove(id).exec();
       return result;
     }
 
     async updateById(id, body) {
-      let model =  this.ctx.model.Task;
+      let model = this.ctx.model.Task;
       const result = await model
         .findByIdAndUpdate(id, body, {
           new: true
@@ -60,7 +60,7 @@ module.exports = app => {
     }
 
     async updateById(id, body) {
-      let model =  this.ctx.model.Task;
+      let model = this.ctx.model.Task;
       const result = await model
         .findByIdAndUpdate(id, body, {
           new: true
@@ -69,8 +69,15 @@ module.exports = app => {
       return result;
     }
 
+
+    async updateVoiceUrl(taskid,relateUrl) {
+      let model = this.ctx.model.Task;
+      await model.updateOne({ _id: taskid }, { voiceUrl: relateUrl }, {});
+      return true;
+    }
+
     async replaceById(id, newDocument) {
-      let model =  this.ctx.model.Task;
+      let model = this.ctx.model.Task;
       await model.findByIdAndRemove(id).exec();
       const result = await model.create(newDocument);
       return result;
